+++ src/pages/Opportunities/index.tsx
@@ -598,9 +598,7 @@ const Opportunities = () => {
                 {
                   icon: <FaPercentage/>,
                   label: "Probability :",
-                  field: "probability"
-                  
-                 
+                  field: "probability" 
                   
                 },
                 {